Improve 2D-histogram intitial shift estimation algorithm (#286)
There may be some extremely small changes in the test results due to this PR but those tests should be OKified @jhunkeler @stsci-hack. The first reason for the difference is that initial shift will have 0.5 pixels difference which can lead to more sources being matched. The second one is that `tolerance` parameter is no longer ignored when `use2dhist` is enabled. This will lead gain to more sources being matched.